requirements.txt
google-generativeai
PyYAML
PyMuPDF
Pillow
python-magic
pandas
